Fix RutabagaVirtioGpu::unref_resource
Make sure that  RutabagaVirtioGpu::unref_resource removes the resource from
the `resources` BTreeMap. Not doing this could cause problems if the driver
attempts to create a new resource with the same id as the previous already
destroyed resource.

Note that I didn't actually observe this happening, but
it is something that should be valid for the driver to do.

Signed-off-by: Matej Hrica <mhrica@redhat.com>